Removed fail catch for the final two cp commands
Removed fail catch for the final two cp commands. On my setup sometimes the script is ran multiple times as part of a build process. I would get errors copying if the src/dest were the same file even though all other operations were successful. 

Since I'm using this repo in my projects I decided to remove the fail catch on the copy to keep the exit status as 0. Plus if the files are the same the cp failing really isn't an error.